+++ CMakeLists.txt
@@ -17,8 +17,8 @@ include_directories(include)
 
 file(GLOB all_srcs "${PROJECT_SOURCE_DIR}/include/*.h"
                    "${PROJECT_SOURCE_DIR}/include/*.hpp"
-        	   "${PROJECT_SOURCE_DIR}/src/*.cpp"
-        	   "${PROJECT_SOURCE_DIR}/src/*.c")	
+        	       "${PROJECT_SOURCE_DIR}/src/*.cpp"
+        	   "${PROJECT_SOURCE_DIR}/src/*.c")
 
 add_executable(${APP_TARGET}
 	       ${all_srcs})
