+++ src/Carbonator.cpp
@@ -1,7 +1,6 @@
 #include "Carbonator.h"
 #include <cmath>
 
-
 using namespace std::chrono ;
